requirements.txt
# Deep Learning
torch>=2.1.0
torchvision>=0.16.0
ultralytics>=8.0.0

# ONNX
onnx>=1.14.0
onnxruntime>=1.16.0

# Computer Vision
opencv-python>=4.8.0
albumentations>=1.3.0
pillow>=10.0.0

# API
fastapi>=0.103.0
uvicorn>=0.23.0
python-multipart>=0.0.6

# Demo
gradio>=4.0.0
plotly>=5.17.0

# Experiment Tracking
wandb>=0.15.0

# Data Processing
numpy>=1.24.0
pandas>=2.0.0
pyyaml>=6.0.0
tqdm>=4.65.0

# Testing
pytest>=7.4.0
pytest-benchmark>=4.0.0

# Utilities
requests>=2.31.0
python-dotenv>=1.0.0
